requirements.txt
streamlit
pandas
plotly
openai
python-dotenv
bcrypt
google-analytics-data
python-docx
